Entities + animations
Stop clearing entity layer:
* prevents sprite flickering
* allows animations to finish
* sometimes leads to sprites 'floating' from place to place
Suspect 'floating' is due to unfinished, conflicting animations running.
Add shark entity.
Rename animal entity to mushroom.
