+++ src/web-components/side-menu.ts
@@ -16,7 +16,7 @@ import BackpackIcon from "../shoelace/assets/icons/backpack4.svg";
 import PersonIcon from "../shoelace/assets/icons/person.svg";
 import HandleIcon from "../shoelace/assets/icons/grip-vertical.svg";
 import { SlIconButton } from "@shoelace-style/shoelace";
-import { CachedSprite } from "../assets";
+import { CachedTexture } from "../assets";
 import { Actor } from "../entities/actor";
 
 export interface MenuTab {
@@ -29,7 +29,7 @@ export type TopLevelMenu = "Entities" | "Cities" | "Resources" | "Build";
 
 export interface MenuItem {
   id: string;
-  icon: CachedSprite;
+  icon: CachedTexture;
   clickHandler?: () => void;
   label?: string;
   tooltip?: string;
